Allow the selector pos to be set programmatically
Thanks to the swap to HSV, it's easy to make it so we can take a color and
then make the pointer go to there. There were also some code cleanups and
fixes, and the colorwheel should actually have all the colors now.

There are magic numbers here that really need to be fixed, though